test: Add initial install test, reuse logic if possible
